refactor(core): clarify planning module structure
