--- packages/core/src/ai-model/service-caller/types.d.ts
@@ -1,1 +1,0 @@
-declare module 'dirty-json';
