+++ packages/core/tests/unit-test/llm-planning.test.ts
@@ -1,10 +1,10 @@
 import { getModelAdapter } from '@/ai-model/models';
 import { descriptionForAction } from '@/ai-model/prompt/llm-planning';
+import { parseXMLPlanningResponse } from '@/ai-model/workflows/planning';
 import {
   parseMarkFinishedIndexes,
   parseSubGoalsFromXML,
-} from '@/ai-model/prompt/util';
-import { parseXMLPlanningResponse } from '@/ai-model/workflows/planning';
+} from '@/ai-model/workflows/planning/standard-planning-parser';
 import { getMidsceneLocationSchema } from '@/common';
 import { buildYamlFlowFromPlans } from '@/common';
 import {
